==> 03 Scripts/01_Data Checks_customers.ipynb <==
# Data Checks on: customers.csv

#### Contents. 
- Missing Data Check
- Duplicate Check
- Mixed-type Data Check
- Other inconsistency, if needed, check
- Summary if changes to the original dataframe

## Importing libraries and data

#importing the libraries

import pandas as pd
import numpy as np
import os

#importing datasets:
path = r'C:\Users\chris\Documents\Instacart Basket Analysis'
df_cust = pd.read_csv(os.path.join(path,'02 Data', 'Original Data', 'customers.csv'))

## Viewing the dataframe

# size:
df_cust.shape

# columns:
df_cust.columns

# looks:
df_cust

# data types:
df_cust.dtypes

# basic stats:
df_cust.describe().round(1)

# basic stats: gender distribution
df_cust.value_counts('Gender')

# basic stats: fam status
df_cust.value_counts('fam_status')

### Missing Data Check

# checking for any missing values:
df_cust.isna().sum()

##### Missing values in first names are ok. We actually don't need the column, since user_id is enough to tell all users apart.

### Duplicated Data Check

# looking for duplicates:
df_dups = df_cust[df_cust.duplicated()]
df_dups

##### No duplicates.

### Mixed-typed Data Check

# checking for mixed-typed data:
for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].applymap(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

# another check: show me data types of all columns
df_cust.dtypes

##### The mixed data type comes from the missing values in 'First Name'. Since we don't need the column, no steps taken here.

## Summary: What needs to be addressed:

##### Dropping column 'First Name' and surname - columns for data privacy
##### changing id columns into strings
##### changing data type to datetime in column 'date_joined'

## Changing and Exporting the dataframe >> customers_wrangled.pkl

# changing ids into string variables:
df_cust['user_id'] = df_cust['user_id'].astype(str)

# changing the data type in 'date_joined' from object to datetime:
df_cust['date_joined'] = pd.to_datetime(df_cust['date_joined'])

# checking the data types:
df_cust.dtypes

# creating a wrangled dataframe, dropping columns 'First Name' and 'Surnam'
df_cust_clean = df_cust.drop(columns = ['First Name','Surnam' ])
df_cust_clean

### Exporting the dataframe

# exporting the dataframe:
df_cust_clean.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'customers_wrangled.pkl'))
