==> 03 Scripts/01_Data Checks_departments.ipynb <==
# Data Checks on: departments.csv

#### Contents. 
- Missing Data Check
- Duplicate Check
- Mixed-type Data Check
- Other inconsistency, if needed, check
- Summary if changes to the original dataframe

## Importing libraries and data

#importing the libraries

import pandas as pd
import numpy as np
import os

#importing datasets:
path = r'C:\Users\chris\Documents\Instacart Basket Analysis'
df_deps = pd.read_csv(os.path.join(path,'02 Data', 'Original Data', 'departments.csv'))

## Viewing the dataframe

# size:
df_deps.shape

# columns:
df_deps.columns

# looks:
df_deps

# data types:
df_deps.dtypes

# basic stats:
df_deps.describe().round(1)

##### This data needs to be transposed. 

### Missing Data Check

# checking for any missing values:
df_deps.isna().sum()

### Duplicated Data Check

# looking for duplicates:
df_dups = df_deps[df_deps.duplicated()]
df_dups

##### Since it's only one row, there are no duplicated rows.

### Mixed-typed Data Check

# checking for mixed-typed data:
for col in df_deps.columns.tolist():
  weird = (df_deps[[col]].applymap(type) != df_deps[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_deps[weird]) > 0:
    print (col)

##### no mix-typed data

## Summary: What needs to be addressed:

#### The data is in a wide format, when a vertical format would be best. I transpose it. It is actually a data dictionary, explaining what those department numbers refer to.

# Changing and Exporting the dataframe >> departments_wrangled.csv

# transposing the whole dataframe:
df_deps.T

# assigning new dataframe to this and calling the df:
df_deps_t=df_deps.T
df_deps_t.head

#adding an index, because we need it in the next steps:
df_deps_t.reset_index()

# create a new header and call it:
new_header = df_deps_t.iloc[0]
new_header

# remove the first row in the dataframe by creating an new dataframe without it:
df_deps_t_new = df_deps_t[1:]
df_deps_t_new

# set the header row as the actual new header:
df_deps_t_new.columns = new_header

# calling the final dataframe:
df_deps_t_new

##### Last check only reading the contents. Data looks fine so far - what could be meant with 21 = 'missing' we'll check out when merging the dataframes.

### Exporting the dataframe

# exporting the dataframe:
df_deps_t_new.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.pkl'))
