==> 03 Scripts/01_Data Checks_orders.ipynb <==
# Data Checks on: orders.csv

#### Contents. 
- Missing Data Check
- Duplicate Check
- Mixed-type Data Check
- Other inconsistency, if needed, check
- Summary if changes to the original dataframe

## Importing libraries and data

# Import libraries
import pandas as pd
import numpy as np
import os

#importing dataset:
path = r'C:\Users\chris\Documents\Instacart Basket Analysis'
df_ords = pd.read_csv(os.path.join(path,'02 Data', 'Original Data', 'orders.csv'))

## Viewing the dataframe

# size:
df_ords.shape

# columns:
df_ords.columns

# looks:
df_ords

# data types:
df_ords.dtypes

# basic stats:
df_ords.describe().round(1)

##### checking on the column 'eval_set' more closely:

# What values are included and how many of each:
df_ords['eval_set'].value_counts(dropna = False)

# Do these two make sense together:
df_ords[['days_since_prior_order', 'eval_set']].value_counts(dropna=False)

### Missing Data Check

# checking for any missing values:
df_ords.isnull().sum()

##### 206.209 cases of missing data - that needs to be examined.

#Creating a subset only with those missing values in 'days_since_prior_order':
df_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]
df_nan

# missing data: seems to be every first order number:
df_nan['order_number'].value_counts(dropna=False)

#### Checking further details with a crosstab:

# creating the crosstab and storing it in the object crosstab (in the code first rows, then columns)
crosstab = pd.crosstab(df_ords['days_since_prior_order'], df_ords['order_number'], dropna = False)

# Copy in a spreadsheet in Excel:
crosstab.to_clipboard()

# Or calling the crosstab here:
crosstab

##### Conclusion: the missing data is logically correct and thus not 'missing': there can be no day since last/prior order before the first order. 

### Duplicated Data Check

# looking for duplicates:
df_dups = df_ords[df_ords.duplicated()]
df_dups

##### No duplicates.

### Mixed-typed Data Check

# checking for mixed-typed data:
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

##### No mixed-typed data

## Summary: What needs to be addressed:

##### 1) column 'eval set' is superfluous. It can be deleted as it only enlargens the size.
##### 2) column 'order_dow' can be renamened into 'order_day_of_week' for better understanding.
##### 3) column 'days_since_prior_order' can be renamened into 'days_since_last_order' for better understanding.
##### 4) column 'order_id' and 'user_id' can be changed into string variables.

## Changing and Exporting the dataframe >> orders_wrangled.pkl

# dropping the column 'eval_set', saving changes in df clean1
df_ords_clean = df_ords.drop('eval_set', axis=1)

# renaming columns:
df_ords_clean.rename(columns = {'order_dow':'order_day_of_week'}, inplace = True)
df_ords_clean.rename(columns = {'days_since_prior_order':'days_since_last_order'}, inplace = True)

# checking the changes: 
df_ords_clean.head()

# order_id and user_id to be changed into string variables:
df_ords_clean[['order_id', 'user_id']] = df_ords[['order_id', 'user_id']].astype('str')

# checking the changes:
df_ords_clean.dtypes

df_ords_clean.shape

# EXPORTING the dataframe into orders_wrangled.pkl
df_ords_clean.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.pkl'))
