==> 03 Scripts/02_merging-all-data.ipynb <==
# Merging all data sets

#### this notebook merges all wrangled data sets: orders, prior, products and customers.
#### the departments data are not essential, they will be used as a data dictionary in the analysis.

## Importing libraries and data

#importing the libraries

import pandas as pd
import numpy as np
import os

#importing datasets:
path = r'C:\Users\chris\Documents\Instacart Basket Analysis'
ords = pd.read_pickle(os.path.join(path,'02 Data', 'Prepared Data', 'orders_wrangled.pkl'))
prods = pd.read_pickle(os.path.join(path,'02 Data', 'Prepared Data', 'products_wrangled.pkl'))
prior = pd.read_pickle(os.path.join(path,'02 Data', 'Prepared Data', 'prior_wrangled.pkl'))
cust = pd.read_pickle(os.path.join(path,'02 Data', 'Prepared Data', 'customers_wrangled.pkl'))

### Planning the merges

#finding KEYs for merge: ords
ords.columns

#finding KEYs for merge: prods
prods.columns

#finding KEYs for merge: prior
prior.columns

#finding KEYs for merge: cust
cust.columns

##### First merge has to be ords with prior (Key = 'order_id') > creating the matching Key with the prods (Key = 'product_id'). Finally we will add the customers (Key = 'user_id').

### 1) Merging orders with prior >> new dataframe: orders_products_combined

#### 1.a) Performing the standard merge, which is an INNER JOIN:

#pre-merge: shape of ords:
ords.shape

#pre-merge: shape of prior:
prior.shape

#merging on: order_id, INNER JOIN
orders_products_combined = ords.merge(prior, on = 'order_id', indicator = True)

#calling the merged dataframe
orders_products_combined

# after merge check: size of merged dataframe:
orders_products_combined.shape

# checking the merge flag:
orders_products_combined['_merge'].value_counts()

#### 1.b) Checking results with the contrary merge, performing an OUTER JOIN:

#merging on: order_id, OUTER JOIN
orders_products_combined_outer = ords.merge(prior, on = 'order_id', indicator = True, how = 'outer')

# checking the merge flag:
orders_products_combined_outer['_merge'].value_counts()

##### The OUTER JOIN shows, that there are more than 200k rows without a match. Reasons for this need to be checked in the next step.

# filter all not merged rows:
unmerged_data = orders_products_combined_outer[orders_products_combined_outer['_merge'] != 'both']

# calling the unmerged_data:
unmerged_data

# apparently th product_id is missing. Subsetting to be sure:
subset_by_product_id_outer = unmerged_data.loc[:, 'product_id']

# if this is the reason, the number of rows equals the number of unmerged data:
subset_by_product_id_outer.shape

# comparative subset in the inner join merge:
subset_by_product_id_inner = orders_products_combined.loc[orders_products_combined['product_id'].isna(), 'product_id']

subset_by_product_id_inner.shape

##### In the orders-dataframe were over 200k rows without a match in the prior_orders-dataframe. In the unmerging rows, essential data is missing, above all the product_id, which will be the Key for the merge with the products dataframe. Therefore, we will continue with the inner joined dataframe.

### 2) Merge 'orders_products_combined' with 'prods' >> new dataframe: orders_products_merged

#### 2.a) Performing the standard merge, which is an INNER JOIN:

#pre-merge: shape of orders_products_combined:
orders_products_combined.shape

#pre-merge: shape of prods:
prods.shape

#merging on: product_id, INNER JOIN
orders_products_merged = orders_products_combined.merge(prods, on = 'product_id')

#calling the merged df:
orders_products_merged

# after merge check: size of merged dataframe:
orders_products_merged.shape

# checking the merge flag:
orders_products_merged['_merge'].value_counts()

#### 2.b) Checking results with the contrary merge, performing an OUTER JOIN:

#merging on: product_id, OUTER JOIN
orders_products_merged_outer = orders_products_combined.merge(prods, on = 'product_id', how = 'outer')

# checking the merge flag:
orders_products_merged_outer['_merge'].value_counts()

##### No unmerged rows here. No further investigations required.

### 3) merging 'orders_products_merged' with 'cust' >> new dataframe: orders_products_all_merge

#### 3.a) Performing the standard merge, which is an INNER JOIN:

#pre-merge: shape of orders_products_merged:
orders_products_merged.shape

#pre-merge: shape of cust
cust.shape

#merging on: user_id, INNER JOIN
orders_products_all_merge = orders_products_merged.merge(cust, on = 'user_id')

# after merge check: size of merged dataframe:
orders_products_all_merge.shape

orders_products_all_merge

# checking the merge flag:
orders_products_all_merge['_merge'].value_counts()

##### 3.b) Checking results with the contrary merge, performing an OUTER JOIN:

#merging on: product_id, OUTER JOIN
orders_products_all_merge_outer = orders_products_merged.merge(cust, on = 'user_id', how = 'outer')

# checking the merge flag:
orders_products_all_merge_outer['_merge'].value_counts()

##### No unmerged rows here. No further investigations required.¶

### 4) check: looking for duplicates after merging

# looking for duplicates:
df_dups = orders_products_all[orders_products_all.duplicated()]
df_dups

##### No duplicates.

### 5) drop the '_merge' column >>> FINAL dataframe: orders_products_all.pkl

# dropping the automatically generated merge-column. We don't need it in the final dataframe.
orders_products_all = orders_products_all_merge.drop('_merge', axis=1)

orders_products_all

orders_products_all.shape

## exporting the final dataframe orders_products_all.pkl

# exporting the merged dataframe as a pkl:
orders_products_all.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_all.pkl'))
