==> 03 Scripts/03_analysis_newcolumns.ipynb <==
# Analysis: new columns

#### This notebook adds more value to the dataframe, by creating new columns and insights. To keep the actual delivery for the client simple, this is done in a separate notebook.¶

### Additional columns interesting for analysis:
- data dictionary for the departments dataframe
- price range of products (low, medium, high price range)
- busy days in the week: when do consumers shop most (least, regularly, most busy)
- busy hours: at what time do customers shop (Fewest, average, most orders)
- customer loyalty, based on max count of orders (new, regular, royal customer)
- Customer Frequency, based on days since last order Non-frequent, Regular, Frequent customer)
- spending flag, based on prices, to distinguish high spending from low spending customers

## Importing libraries and data

#importing the libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

#importing dataset:
path = r'C:\Users\chris\Documents\Instacart Basket Analysis'
orders_products_all = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_all.pkl'))
deps = pd.read_pickle(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.pkl'))

#calling main dataframe to check if imported completely:
orders_products_all.head(3)

### Data Dictionary for convenient analysis

##### this isn't exactly a new column, but a new feature for the analysis and is thus tested here.

# defining the data dictionary:
data_dict = deps.to_dict('index')

# calling the df:
data_dict

# calling: get me department 19, should be 'snacks'
print(data_dict.get('19'))

### new column: price range

# Loc()method, condition 1: price > 15 = 'high-range product'
orders_products_all.loc[orders_products_all['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Loc()method, condition 2:
orders_products_all.loc[(orders_products_all['prices'] <= 15) & (orders_products_all['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

# Loc()method, condition 3:
orders_products_all.loc[orders_products_all['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# check new column:
orders_products_all['price_range_loc'].value_counts(dropna = False)

# multiple aggregations: mean, min, max
orders_products_all.groupby('price_range_loc').agg({'prices': ['mean', 'min', 'max']}).round(1)

### new column: busy days in the week:

# First, find out what is the busiest day of the week:
orders_products_all['order_day_of_week'].value_counts(dropna=False)

# For-Loop to create new column with labels “Busiest day,” “Least busy,” and “Regularly busy”:
result = []

for value in orders_products_all["order_day_of_week"]:
  if value == 0:
    result.append("Most busy")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

#checking if it worked:
print(result[:10])

# inserting "result" as a new column:
orders_products_all['busiest_day'] = result

# count values in new column:
orders_products_all['busiest_day'].value_counts(dropna = False)

#show df with new column
orders_products_all.head(5)

### new column: busy days in the week (2 days)

# first, check again on busiest days. For more convienience, sort desc:
orders_products_all['order_day_of_week'].value_counts(dropna=False).sort_values(ascending=False)

# days are identified, so create the loop, using a new result list (result2):
result2 = []

for value in orders_products_all["order_day_of_week"]:
    if value in (0, 1):
        result2.append("Most busy")
    elif value in (3, 4):
        result2.append("Least busy")
    else:
        result2.append("Regularly busy")

#checking if it worked:
print(result2[:10])

# inserting "result2" as a new column:
orders_products_all['busiest_days'] = result2

# checking the extended dataframe:
orders_products_all.head(5)

# count values in new column:
orders_products_all['busiest_days'].value_counts(dropna = False)

### new column: busies hours of the day

# first, check busiest hours:
orders_products_all['order_hour_of_day'].value_counts(dropna=False).sort_values(ascending=False)

# busiest periods of hours would be, starting at midnight:
## 0 - 6: Fewest orders
## 7 - 9: Average orders
## 10 - 17: Most orders
## 18 - 22: Average orders
## 23: Fewest orders
### for-loop can be as follows:
result3 = []

for value in orders_products_all["order_hour_of_day"]:
    if value in (10,11,12,13,14,15,16,17):
        result3.append("Most orders")
    elif value in (7,8,9,18,19,20,21,22):
        result3.append("Average orders")
    elif value in (0,1,2,3,4,5,6,23):
        result3.append("Fewest orders")
    else:
        result3.append("check again!")

# inserting "result3" as a new column:
orders_products_all['busiest_period_of_day'] = result3

# checking the extended dataframe:
orders_products_all.head(5)

# count values in new column - if done correctly, the else-value 'check again!' must not appear:
orders_products_all['busiest_period_of_day'].value_counts(dropna = False)

## customer loyalty

### 1) COUNT of orders > MAX(COUNT) in new column: max_order

#Definition of customer loyalty:

## If max orders the user has made is over 40, then the customer will be labeled a “Loyal customer.”
## If max orders the user has made is over 10 but less than or equal to 40, then the customer will be labeled a “Regular customer.”
### If max orders the user has made is less than or equal to 10, then the customer will be labeled a “New customer.”

# creating a new column 'max order'
orders_products_all['max_order'] = orders_products_all.groupby(['user_id'])['order_number'].transform(np.max)

# as that created a new column, we're heading() the frame to check:
orders_products_all.head()

### 2) flagging the loyalty based on number of max orders:

# Flagging the loyal customers: Deriving Columns with loc()
# loyal customer:
orders_products_all.loc[orders_products_all['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

# regular customer:
orders_products_all.loc[(orders_products_all['max_order'] <= 40) & (orders_products_all['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

# new customer:
orders_products_all.loc[orders_products_all['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

# checking on flags, no NaN should appear:
orders_products_all['loyalty_flag'].value_counts(dropna=False)

# checking the new flag in the dataframe:
orders_products_all.head(5)

### Customer Frequency

# Frequency labels and definition:
## If mean ('mean_days_since_last_order') > 20, then the customer will be labeled a “Non-frequent customer”
## If mean ('mean_days_since_last_order') > 10 AND <=20 , then the customer will be labeled a “Regular customer”
## If mean ('mean_days_since_last_order') <=10, then the customer will be labeled a “Frequent customer”

# first, we need a column with mean days_since_prior_order, just as we did with the 'max_order':
orders_products_all['mean_days_since_last_order'] = orders_products_all.groupby(['user_id'])['days_since_last_order'].transform(np.mean)

# as that created a new column 'mean_days_since_last_order', we're heading() the frame to check:
orders_products_all.head(5)

# assigning the labels now:
## Non-frequent customer: > 20
orders_products_all.loc[orders_products_all['mean_days_since_last_order'] > 20, 'frequency_type'] = 'Non-frequent customer'

# assigning the labels now:
## Regular customer: > 10 AND <=20 
orders_products_all.loc[(orders_products_all['mean_days_since_last_order'] > 10) & (orders_products_all['mean_days_since_last_order'] <=20), 'frequency_type'] = 'Regular customer'

# assigning the labels now:
## low spender: <=10
orders_products_all.loc[orders_products_all['mean_days_since_last_order'] <=10, 'frequency_type'] = 'Frequent customer'

# checking on flags, no NaN should appear:
orders_products_all['frequency_type'].value_counts(dropna=False)

# Subsetting the NaN's:
frequency_type_nan = orders_products_all.loc[orders_products_all['frequency_type'].isnull()]

# calling my subset
frequency_type_nan

##### The NaN's are ol: It's customers with 'NaN' in the column 'days_since_last_order'. We can ignore that. The aggregation and labeling is correct.

### new column: spending flag

#These are the labels:
## If mean ('prices') < 10, then the customer will be labeled a “low spender”
## If mean ('prices') >= 10, then the customer will be labeled a “high spender”

# first, we need a column with mean prices, just as we did with the 'max_order':
orders_products_all['mean_prices'] = orders_products_all.groupby(['user_id'])['prices'].transform(np.mean)

# as that created a new column 'mean_prices', we're heading() the frame to check:
orders_products_all.head(5)

# assigning the labels now:
## low spender:
orders_products_all.loc[orders_products_all['mean_prices'] < 10, 'spender_type'] = 'low spender'

# assigning the labels now:
## high spender:
orders_products_all.loc[orders_products_all['mean_prices'] >=10, 'spender_type'] = 'high spender'

# checking on flags, no NaN should appear:
orders_products_all['spender_type'].value_counts(dropna=False)

## Exporting the enhanced dataframe

# exporting the extended dataframe as a pkl:
orders_products_all.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_all.pkl'))
